refactor: move use cases (the business logic) to AMW_business
