+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/property/boundary/AddTagUseCase.java
@@ -1,4 +1,4 @@
-package ch.mobi.itc.mobiliar.rest.tags.usecases;
+package ch.puzzle.itc.mobiliar.business.property.boundary;
 
 import ch.puzzle.itc.mobiliar.business.property.control.PropertyTagEditingService;
 import ch.puzzle.itc.mobiliar.business.property.entity.PropertyTagEntity;
