+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/property/boundary/TagCommand.java
@@ -1,6 +1,5 @@
-package ch.mobi.itc.mobiliar.rest.tags.usecases;
+package ch.puzzle.itc.mobiliar.business.property.boundary;
 
-import ch.mobi.itc.mobiliar.rest.dtos.TagDTO;
 import ch.puzzle.itc.mobiliar.common.exception.ValidationException;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
@@ -13,17 +12,12 @@ public class TagCommand {
     @NotNull
     private String name;
 
-    public TagCommand(TagDTO tagDTO) throws ValidationException {
-        this.requireNonNull(tagDTO);
-        this.requireNotEmpty(tagDTO.getName());
-        this.name = tagDTO.getName();
+    public TagCommand(String tagName) throws ValidationException {
+        this.requireNotEmpty(tagName);
+        this.name = tagName;
     }
 
     private void requireNotEmpty(String string) throws ValidationException {
         if (StringUtils.isBlank(string)) throw new ValidationException("Tag name must not be null or empty.");
-    }
-
-    private void requireNonNull(Object obj) throws ValidationException {
-        if (obj == null) throw new ValidationException("Tag must not be null.");
     }
 }
