+++ AMW_rest/pom.xml
@@ -175,27 +175,5 @@
 			<version>1.18.6</version>
 			<scope>test</scope>
 		</dependency>
-
-		<!-- test jar dependency on AMW_business - mostly to reuse the PersistenceTestRunner -->
-		<dependency>
-			<groupId>ch.puzzle.itc.mobiliar</groupId>
-			<artifactId>AMW_business</artifactId>
-			<type>test-jar</type>
-			<version>${parent.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.hibernate</groupId>
-			<artifactId>hibernate-entitymanager</artifactId>
-			<version>5.1.17.Final</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.h2database</groupId>
-			<artifactId>h2</artifactId>
-			<!-- downgraded from 1.4.197 see https://github.com/h2database/h2database/issues/1073 -->
-			<version>1.4.196</version>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
 </project>
