+++ AMW_rest/src/main/java/ch/mobi/itc/mobiliar/rest/tags/boundary/TagsRest.java
@@ -1,10 +1,10 @@
 package ch.mobi.itc.mobiliar.rest.tags.boundary;
 
 import ch.mobi.itc.mobiliar.rest.dtos.TagDTO;
-import ch.mobi.itc.mobiliar.rest.tags.usecases.AddTagUseCase;
-import ch.mobi.itc.mobiliar.rest.tags.usecases.ListTagsUseCase;
-import ch.mobi.itc.mobiliar.rest.tags.usecases.RemoveTagUseCase;
-import ch.mobi.itc.mobiliar.rest.tags.usecases.TagCommand;
+import ch.puzzle.itc.mobiliar.business.property.boundary.AddTagUseCase;
+import ch.puzzle.itc.mobiliar.business.property.boundary.ListTagsUseCase;
+import ch.puzzle.itc.mobiliar.business.property.boundary.RemoveTagUseCase;
+import ch.puzzle.itc.mobiliar.business.property.boundary.TagCommand;
 import ch.puzzle.itc.mobiliar.business.property.entity.PropertyTagEntity;
 import ch.puzzle.itc.mobiliar.common.exception.ValidationException;
 import io.swagger.annotations.Api;
@@ -49,7 +49,7 @@ public class TagsRest {
     @POST
     @ApiOperation(value = "Adds one tag")
     public Response addOneTag(TagDTO tagDTO) throws ValidationException {
-        TagCommand tagCommand = new TagCommand(tagDTO);
+        TagCommand tagCommand = new TagCommand(tagDTO.getName());
         PropertyTagEntity newTag = addTagUseCase.addTag(tagCommand);
         return Response
                 .status(CREATED)
